addressbook, device_id, user_agent services
